code/binary-tree/ViewModel/src/test/java/ru/unn/agile/binarytree/viewmodel/ViewModelTests.java
package ru.unn.agile.binarytree.viewmodel;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class ViewModelTests {
    private ViewModel viewModel;

    @Before
    public void setUp() {
        viewModel = new ViewModel();
    }

    @After
    public void tearDown() {
        viewModel = null;
    }

    @Test
    public void canSetDefaultValues() {
        assertEquals("", viewModel.userKeyProperty().get());
        assertEquals("", viewModel.operationResultProperty().get());
        assertEquals(Operation.ADD, viewModel.operationProperty().get());
    }
}
